% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadoe.R
\name{get_school_discipline_data}
\alias{get_school_discipline_data}
\title{Get school discipline data from the CA DOE}
\usage{
get_school_discipline_data(year, category)
}
\arguments{
\item{year}{String, in '1718' format}

\item{category}{Either 'susp' for suspensions or 'exp' for expulsions}
}
\value{
a large data frame
}
\description{
Get school discipline data from the CA DOE
}
\examples{
\dontrun{
suspensions1718 <- get_school_discipline_data("1718", "susp")
}
}
