% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadoe.R
\name{get_school_enr_data}
\alias{get_school_enr_data}
\title{Download Enrollment Data by Year from the CA DOE}
\usage{
get_school_enr_data(year)
}
\arguments{
\item{year}{String, in '2018-19' format}
}
\value{
a data frame
}
\description{
Download Enrollment Data by Year from the CA DOE
}
\examples{
\dontrun{
enr178 <- get_school_enr_data("2017-18")
}
}
