% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadoe.R
\name{get_school_grad_data}
\alias{get_school_grad_data}
\title{Get graduate data from the CA DOE}
\usage{
get_school_grad_data(year, category)
}
\arguments{
\item{year}{String, in '2017-18' format}

\item{category}{String, either 'Graduates', 'GradEth', 'Dropouts', or 'UCGradEth'}
}
\value{
a large data frame
}
\description{
Get graduate data from the CA DOE
}
\examples{
\dontrun{
ucgrads178 <- get_school_grad_data("2017-18", "UCGradEth")
}
}
